feat: update README + rename tokenInfos folder to infos
