+++ src/packages/tokens/.gitignore
@@ -1,2 +1,2 @@
 # Automatically generated PancakeSwap token list
-/tokenInfos/pancakeSwapTokens/bscMainnet.ts
+/infos/pancakeSwapTokens/bscMainnet.ts
