+++ src/packages/tokens/infos/pancakeSwapTokens/index.tsx
