+++ src/packages/tokens/utilities/getTokens/index.ts
@@ -1,6 +1,6 @@
 import { ChainId } from 'types';
 
-import { tokens } from '../../tokenInfos/commonTokens';
+import { tokens } from '../../infos/commonTokens';
 
 export interface GetTokensInput {
   chainId: ChainId;
